+++ uefi-test-runner/src/proto/pi/mod.rs
@@ -1,9 +1,9 @@
 use uefi::prelude::*;
 
-pub fn test(bt: &BootServices) {
+pub fn test(image: Handle, bt: &BootServices) {
     info!("Testing Platform Initialization protocols");
 
-    mp::test(bt);
+    mp::test(image, bt);
 }
 
 mod mp;