+++ uefi-test-runner/src/proto/rng.rs
@@ -5,11 +5,7 @@ use uefi::table::boot::{BootServices, OpenProtocolAttributes, OpenProtocolParams
 pub fn test(image: Handle, bt: &BootServices) {
     info!("Running rng protocol test");
 
-    let handle = *bt
-        .find_handles::<Rng>()
-        .expect("Failed to get Rng handles")
-        .first()
-        .expect("No Rng handles");
+    let handle = bt.get_handle_for_protocol::<Rng>().expect("No Rng handles");
 
     let mut rng = bt
         .open_protocol::<Rng>(